+++ public/index.html
@@ -2,18 +2,17 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/logo.png" />
+    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#ffffff" />
+    <meta name="theme-color" content="#FDF9F2" /> 
     <meta
       name="description"
-      content="A meal-based website that gives you all the information you need about your favourite breakfast, lunch and dinner. built with react-typescript & redux."
+      content="A cuisine-themed website. search for your favorite meals and ingredients. Built by Okoye Charles and Ernest Adonu."
     />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo.png" />
     <title>Lavish Cuisine</title>
   </head>
   <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
   </body>
 </html>
