+++ src/styles/Meal.css
@@ -1,5 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Dancing+Script&family=Kalam:wght@300;400;700&display=swap');
-
 .meal {
   padding-bottom: 3rem;
 }
