+++ src/styles/PageNotFound.css
@@ -1,69 +1,19 @@
 .pageNotFound {
-  padding: 0.5em 1em;
-}
+  height: calc(100vh - var(--nav-height));
+  display: grid;
+  place-items: center;
+  place-content: center;
+  padding: 1rem;
+  color: var(--color-brown);
 
-.pageNotFound__header {
-  display: flex;
-  flex-direction: column;
-  border-bottom: 0.5em solid #c0841d;
-  margin-bottom: 1em;
-}
+  svg {
+    font-size: 10rem;
+    color: var(--color-milk-brown);
+  }
 
-.pageNotFound__header > h1 {
-  color: #51380b;
-  font-weight: 300;
-  text-align: center;
-}
+  h1 {
+    font-size: 1.15rem;
+    text-align: center;
 
-.pageNotFound__icon {
-  font-size: 12.5em;
-  margin-inline: auto;
-  color: #c0841d;
-}
-
-.pageNotFound__content > h2 {
-  font-weight: 400;
-  font-size: 1.2em;
-}
-
-.pageNotFound__content > ul {
-  list-style-type: none;
-  padding: 0;
-}
-
-.pageNotFound__content > ul > li {
-  font-weight: 300;
-  margin-bottom: .5em;
-}
-
-.pageNotFound__content > ul div {
-  color: #8f5b01;
-  width: fit-content;
-}
-
-.pageNotFound__content > ul div:hover {
-  cursor: pointer;
-  color: #916417;
-  text-decoration: underline;
-}
-
-.pageNotFound__content > ul div:hover + span {
-  background: #c68919e0;
-}
-
-.pageNotFound__content > ul span {
-  background: #c68919a1;
-  color: white;
-  padding: .25em;
-}
-
-.pageNotFound__content > ul span::before {
-  content: '↳ ';
-  color: white;
-}
-
-@media screen and (max-width: 700px) {
-  .pageNotFound__content > ul span {
-    background: #c68919e0;
   }
 }
